upd VO folder
